Move all steps/pipelines to appropriate folders.
