+++ examples/slack_alert/steps/__init__.py
@@ -1,3 +1,4 @@
 from .evaluator import evaluator
+from .formatter import test_acc_ask_formatter, test_acc_post_formatter
 from .importer import importer
 from .trainer import svc_trainer, svc_trainer_mlflow
